% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libsassr-package.R
\docType{package}
\name{libsassr-package}
\alias{libsassr}
\alias{libsassr-package}
\title{libsassr: Wraps the LibSass C++ Library}
\description{
Wraps the LibSass C++ Library in R.
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
