% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{sass_compile_file}
\alias{sass_compile_file}
\title{Compile Sass source code (from a file)}
\usage{
sass_compile_file(filename, outfile = "", output_style = "nested",
  precision = 5)
}
\arguments{
\item{filename}{File path containing Sass source code to compile.}

\item{outfile}{Path to store resulting CSS. Default is \code{""} and just prints
in standard output.}

\item{output_style}{An optional coding style of the compiled result. Choose
one of: \code{"nested"} (default), \code{"expanded"}, \code{"compact"}, \code{"compressed"}.}

\item{precision}{Optional precision for numbers. Default is \code{5}.}
}
\description{
Compile Sass source code (from a file)
}
\examples{
string <- "a { b { color: blue; } }"
write(string, tmp <- tempfile())
sass_compile_file(paste0(tmp, 1))
}
