% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{sass_compile_string}
\alias{sass_compile_string}
\title{Compile Sass source code (from a string)}
\usage{
sass_compile_string(string, outfile = "", output_style = "nested",
  precision = 5)
}
\arguments{
\item{string}{Sass source code (as a string) to compile.}

\item{outfile}{Path to store resulting CSS. Default is \code{""} and just prints
in standard output.}

\item{output_style}{An optional coding style of the compiled result. Choose
one of: \code{"nested"} (default), \code{"expanded"}, \code{"compact"}, \code{"compressed"}.}

\item{precision}{Optional precision for numbers. Default is \code{5}.}
}
\description{
Compile Sass source code (from a string)
}
\examples{
string <- "a { b { color: blue; } }"
sass_compile_string(string)
}
